src/etl/requirements.txt
requests>=2.31.0,<3.0.0
pydantic>=2.6.0,<3.0.0
pandas>=2.2.0,<3.0.0
pyarrow>=15.0.0,<20.0.0
pyyaml>=6.0,<7.0
google-cloud-bigquery>=3.17.0,<4.0.0
google-cloud-storage>=2.14.0,<3.0.0
